- Добавлены тесты

+++ 2Sem.HomeWork4/2.4.1/2.4.1/Addition.cs
@@ -4,8 +4,14 @@ using System.Text;
 
 namespace _2._4._1
 {
+    /// <summary>
+    /// Class Addition
+    /// </summary>
     public class Addition : Operation
     {
+        /// <summary>
+        /// Output expression with addition
+        /// </summary>
         public override void Print()
         {
             Console.Write("(");
@@ -15,7 +21,11 @@ namespace _2._4._1
             Console.Write(")");
         }
 
-        public override int Counting()
+        /// <summary>
+        /// Addition counting
+        /// </summary>
+        /// <returns>Result of addition</returns>
+        public override double Counting()
             => LeftChild.Counting() + RightChild.Counting();
     }
 }
