+++ 2Sem.HomeWork4/2.4.1/2.4.1/Multiplication.cs
@@ -4,8 +4,14 @@ using System.Text;
 
 namespace _2._4._1
 {
+    /// <summary>
+    /// Multiplication class
+    /// </summary>
     public class Multiplication : Operation
     {
+        /// <summary>
+        /// Multiplication output
+        /// </summary>
         public override void Print()
         {
             Console.Write("(");
@@ -15,7 +21,11 @@ namespace _2._4._1
             Console.Write(")");
         }
 
-        public override int Counting()
+        /// <summary>
+        /// Result of multiplication
+        /// </summary>
+        /// <returns></returns>
+        public override double Counting()
             => LeftChild.Counting() * RightChild.Counting();
     }
 }
