+++ 2Sem.HomeWork4/2.4.1/2.4.1/Number.cs
@@ -4,20 +4,31 @@ using System.Text;
 
 namespace _2._4._1
 {
+    /// <summary>
+    /// Tree element
+    /// </summary>
     public class Number : INode
     {
-        public int value { get; set; }
+        public double value { get; set; }
 
-        public Number(int value)
+        public Number(double value)
         {
             this.value = value;
         }
+
+        /// <summary>
+        /// Value output
+        /// </summary>
         public void Print()
         {
             Console.Write(value);
         }
 
-        public int Counting() => value;
+        /// <summary>
+        /// Return value of tree element
+        /// </summary>
+        /// <returns>value</returns>
+        public double Counting() => value;
         
     }
 }
