+++ 2Sem.HomeWork4/2.4.1/2.4.1/Subtraction.cs
@@ -4,8 +4,14 @@ using System.Text;
 
 namespace _2._4._1
 {
+    /// <summary>
+    /// Subtraction's class
+    /// </summary>
     public class Subtraction : Operation
     {
+        /// <summary>
+        /// Subtraction's output
+        /// </summary>
         public override void Print()
         {
             Console.Write("(");
@@ -15,7 +21,11 @@ namespace _2._4._1
             Console.Write(")");
         }
 
-        public override int Counting()
+        /// <summary>
+        /// Subtraction's counting
+        /// </summary>
+        /// <returns>Subtraction's result</returns>
+        public override double Counting()
             => LeftChild.Counting() - RightChild.Counting();
     }
 }
